Preliminary command framework shenanigans
