+++ package.json
@@ -17,7 +17,7 @@
     "typescript": "^5.3.3"
   },
   "dependencies": {
-    "@b01lers/ctfd-api": "^1.2.0",
+    "@b01lers/ctfd-api": "^1.2.1",
     "@slack/bolt": "^4.2.1",
     "adm-zip": "^0.5.12",
     "async-lock": "^1.4.1",
